requirements.txt
langchain~=0.3.27
langgraph~=0.6.6
langchain-core~=0.3.75
langchain-community
ipykernel
python-dotenv~=1.1.1
langchain-openai
langchain-groq
pydantic~=2.11.7
pandas
openpyxl
requests~=2.32.5
fastapi
uvicorn